neuralnetwork/requirements.txt
# SkillSevak - AI-Powered Resume Matching System
# Requirements File - Phase 1

# =============================================================================
# CORE DJANGO
# =============================================================================
Django==5.2
python-decouple==3.8
gunicorn==21.2.0

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# =============================================================================
# ML/NLP - Dynamic Skill Extraction (NO HARDCODED SKILLS!)
# =============================================================================
spacy==3.7.2
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.2

# =============================================================================
# RESUME PARSING
# =============================================================================
pdfplumber==0.10.3
python-docx==1.1.0

# =============================================================================
# BACKGROUND TASKS (Phase 6)
# =============================================================================
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# =============================================================================
# DATA PROCESSING
# =============================================================================
numpy==1.26.2
pandas==2.1.3

# =============================================================================
# EXTERNAL DATASET LOADING (HuggingFace)
# =============================================================================
datasets>=2.14.0

# =============================================================================
# AUTHENTICATION (Google OAuth)
# =============================================================================
django-allauth==0.58.2

# =============================================================================
# UTILITIES
# =============================================================================
Pillow==10.1.0

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
